Switch to `PHP_CodeSniffer` v3.8.0
Unfortunately, `quizlabs/php_codesniffer` is officially abandoned.
A fork has been created at `phpcsstandards/php_codesniffer`.

As such, we now require `PHP_CodeSniffer` v3.8.0 or later. This
version fixes a lot of things and also adds support for PHP 8.3.

The default PHP version for the action is now PHP 8.3. This does
not impact downstream dependencies of this action as long as they
properly configured the version constraints.